Add support for varied voxels alongside animations

+++ client/event/glfw_framebuffer_size.hh
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: Zlib
-// Copyright (c) 2024, Voxelius Contributors
+// Copyright (C) 2024, Voxelius Contributors
 #ifndef CLIENT_EVENT_GLFW_WINDOW_RESIZE_HH
 #define CLIENT_EVENT_GLFW_WINDOW_RESIZE_HH
 
 struct GlfwFramebufferSizeEvent final {
     int width {};
     int height {};
-    double aspect {};
+    float aspect {};
 };
 
 #endif /* CLIENT_EVENT_GLFW_WINDOW_RESIZE_HH */
