+++ client/event/glfw_scroll.hh
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: Zlib
-// Copyright (c) 2024, Voxelius Contributors
+// Copyright (C) 2024, Voxelius Contributors
 #ifndef CLIENT_EVENT_GLFW_MOUSE_SCROLL_HH
 #define CLIENT_EVENT_GLFW_MOUSE_SCROLL_HH
 
 struct GlfwScrollEvent final {
-    double dx {};
-    double dy {};
+    float dx {};
+    float dy {};
 };
 
 #endif /* CLIENT_EVENT_GLFW_MOUSE_SCROLL_HH */
