+++ shared/inertial.hh
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: Zlib
-// Copyright (c) 2024, Voxelius Contributors
+// Copyright (C) 2024, Voxelius Contributors
 #ifndef SHARED_INERTIAL_HH
 #define SHARED_INERTIAL_HH
 
 namespace inertial
 {
-void update(double frametime);
+void update(float frametime);
 } // namespace inertial
 
 #endif /* SHARED_INERTIAL_HH */
